See if this fixes our 'end of file reached' issue.
We believe that setting clock to localtime means that every VM hwclock
is then set to the localtime of the server (UTC+1 when we're in BST).
This means our VMs start with their hwclock 1 hour ahead of where it
should be, plus its timezone (UTC+1), meaning that it's 2 hours ahead of
UTC, and 1 hour ahead of the actual localtime.

Hopefully setting this to utc will mean that the VM hwclocks are all in
UTC and their timezone settings mean that the servers always have approx
the correct time on start.

Were then hoping that his stops our 'end of file reached' issue as the
puppet pool code will correctly remove expired connections from the
pool. Unlike now, where we believe nt resetting the clock part way
through our puppet run means that connections are allowed to stay around
for much longer after they should have been expired.