Hopefully fix the DNS allocation problem again
When updating the checks to use getaddresses and getnames, I forgot to
take into account that they now worked differently. So I have introduced
the same solution as i've used in other parts of the code to revert back
to getaddress and get name if the results of the other versions return
an empty array.

Plus a test to catch this change in the future.