aparker: Control run of initial boot script with a lockfile
I found the intial boot sequence very confusing. I could see in syslog
that commands were run in rc.local, but when I looked at it it only
contained "exit 0". Until I found the original source I didn't know what
was happening. By overwriting rc.local after it runs, it does provide a
way of having it happen only once, but it also makes any boot failures
harder to diagnose.

This uses a lockfile (/var/lib/firstboot) that is created at the end of
the run and checked at the beginning. If it exists, the rest of the
rc.local script is skipped. Now someone will be able to see what was run
at first boot. This also opens the option of deleting firstboot and
running it again to work on changes to it.