aparker: use sntp to force the time to be set
Back in 2014 the ntpdate command was changed to use -d to get more
output from it. However, -d (debug) also changes it to not set the local
clock. This showed up as a test shortly after boot that checks the time
and failed with errors like:
	02:44:54   1) machine.st-nat-001 behaves like nrpe check_ntp
	02:44:54      Failure/Error: Unable to find matching line from backtrace
	02:44:54      RuntimeError:
	02:44:54        NTP WARNING: Offset 0.5717402697 secs|offset=0.571740s;0.500000;1.000000;
	02:44:54      Shared Example Group: "nrpe" called from /usr/local/lib/site_ruby/timgroup/stackbuilder/stacks/rake/tasks.rb:594
	02:44:54      # /usr/local/lib/site_ruby/timgroup/stackbuilder/stacks/rake/../stacktests/machine/nrpe.rb:17
	02:44:54      # /usr/local/lib/site_ruby/timgroup/stackbuilder/stacks/rake/tasks.rb:598
	02:44:54      # /usr/local/lib/site_ruby/timgroup/stackbuilder/stacks/rake/tasks.rb:88:in `call'
	02:44:54      # /usr/local/lib/site_ruby/timgroup/stackbuilder/stacks/rake/tasks.rb:88:in `sbtask'

After removing the -d flag, the time still wasn't set. According to the
documentation the -b flag should have caused ntpdate to immediately set
the time. A user on askubuntu was in a similar situation. The only thing
that seemed to work was using sntp.

After a few tests it seems to do what we need.

This changes the rc.local firstboot script to use sntp.