Work towards giving us a way to migrate VMs between KVM hosts on different Ubuntu versions
Have KVM hosts use their default machine type for machine. On all versions
of Ubuntu with even vaguely recent versions of the qemu-kvm package
this now does something sensible and defaults to a machine type that's
defined on at least the following LTS version of Ubuntu.

Precise defaults to pc-1.0-qemu-kvm which is supported by Trusty
Trusty defaults to pc-i440fx-trusty which is supported right up to the
  latest non LTS that begins with A.

This does still mean that VMs from a precise KVM host that are migrated
to a Trusty KVM host then can't be migrated to a KVM host > Trusty.
Without us changing the machine and power cycling the VM.

This is good enough for the time being as it begins to offer us live
migration which has always evaded us previously.