Specific the minimum available CPU, instead of taking some weird mixture of host
I'm unclear to what KVM defaults to, however it appears to be some
ancient Pentium II-like CPU, yet with the addition of the "abm" flag. We
have machines without abm, so this is bad as it prevents migrations to
older hardware.

This commit brings us up to the minimum allowed spec, which is a G7
running Precise. This thus adds the following additional flags:
  - vme
  - pat
  - ss
  - pdpe1gb
  - rdtscp
  - constant_tsc
  - ssse3
  - sse4_1
  - sse4_2
  - aes