Change from a magical IP to a name
Commit 7d2557 changed from using a name to an IP in an effort to make
the ntp call more reliable. Unfortunately it make it more prone to
failure because of a changing network. Taking it back to using a name
and hopefully it won't fail very often. Will have to revisit of this
makes it worse.